+++ Arc/src/Platform/OpenGL/OpenGLTextureCubemap.cpp
@@ -182,8 +182,8 @@ namespace ArcEngine
 		for (unsigned int mip = 0; mip < maxMipLevels; ++mip)
 		{
 			// reisze framebuffer according to mip-level size.
-			unsigned int mipWidth  = radianceMapSize * glm::pow(0.5, mip);
-			unsigned int mipHeight = radianceMapSize * glm::pow(0.5, mip);
+			uint32_t mipWidth  = (uint32_t)((float)radianceMapSize * glm::pow(0.5, mip));
+			uint32_t mipHeight = (uint32_t)((float)radianceMapSize * glm::pow(0.5, mip));
 			glBindRenderbuffer(GL_RENDERBUFFER, captureRBO);
 			glRenderbufferStorage(GL_RENDERBUFFER, GL_DEPTH_COMPONENT24, mipWidth, mipHeight);
 			glViewport(0, 0, mipWidth, mipHeight);
